Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["mohamedkhedr <mohamed.khedr@bld.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
numpy = "^1.24.2"
opencv-python = "^4.7.0.72"
pillow = "^9.4.0"
ultralytics = "^8.0.45"
translate = "^3.6.1"
python-translator = "^1.0.3"
beautifulsoup4 = "^4.11.2"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
